fix: corrections diverses + ajout status.go
